requirements.txt
# TitanovaX Trading System Requirements
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Machine Learning
xgboost>=1.7.0
onnxruntime>=1.12.0
onnxmltools>=1.11.0

# Deep Learning (for Informer/TFT)
torch>=1.12.0
transformers>=4.21.0

# Reinforcement Learning
stable-baselines3>=1.6.0
gym>=0.21.0

# API and Web
fastapi>=0.68.0
uvicorn[standard]>=0.18.0
websockets>=10.0
python-telegram-bot>=13.0
requests>=2.28.0

# Data Processing
talib-binary>=0.4.19
pyarrow>=8.0.0
faiss-cpu>=1.7.0

# NLP and Embeddings
sentence-transformers>=2.2.0
pillow>=9.0.0

# Development and Utilities
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
jupyter>=1.0.0

# Async and Performance
aiohttp>=3.8.0
psutil>=5.9.0

# Trading Specific Libraries
MetaTrader5>=5.0.45
ccxt>=4.2.18
yfinance>=0.2.28
pandas-ta>=0.3.14b0

# Infrastructure
redis>=5.0.1
celery>=5.3.4

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pymongo>=4.6.1

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0

# Configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
schedule>=1.2.0
python-dateutil>=2.8.2
pytz>=2023.3.post1

# Security
cryptography>=41.0.8

# Optional: GPU Support
# torch>=1.12.0+cu116  # For CUDA support
# onnxruntime-gpu>=1.12.0  # For GPU inference
